+++ go.mod
@@ -56,3 +56,5 @@ require (
 	google.golang.org/grpc v1.19.0
 	gopkg.in/yaml.v2 v2.2.2
 )
+
+go 1.13
